refactor: split backend logic
